apps/api/pyproject.toml
[project]
name = "hhps-api"
version = "0.1.0"
requires-python = ">=3.12,<3.13"
dependencies = [
  "django==5.0.*",
  "djangorestframework==3.15.*",
  "djangorestframework-simplejwt==5.3.*",
  "psycopg[binary]==3.2.*",
  "python-dotenv==1.0.*",
  "redis==5.0.*",
  "celery==5.4.*",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.*",
  "pytest-django==4.9.*",
  "ruff==0.6.*",
  "mypy==1.11.*",
  "django-stubs==5.0.*",
  "djangorestframework-stubs==3.15.*",
]

[tool.uv]
# Run-from-source Django project, not distributed as a wheel.
package = false

[tool.ruff]
line-length = 100
target-version = "py312"
exclude = ["migrations", "node_modules", ".venv", "**/migrations/**"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "DJ"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["hhps", "tenancy", "accounts", "core", "seed"]

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main"]
exclude = ["migrations/", "\\.venv/"]
ignore_missing_imports = true
# Loose mode: catch obvious bugs, don't fight dynamic Django patterns.
check_untyped_defs = true
disallow_untyped_defs = false
warn_unused_ignores = true
warn_redundant_casts = true

[tool.django-stubs]
django_settings_module = "hhps.settings"
